Move check_eval_commits.dart and test_utils.dart to evals/tool/
