+++ packages/camera/camera_android_camerax/.agents/skills/pre-push-skill/SKILL.md
@@ -48,7 +48,7 @@ Ensure that no commits in this branch were generated by evaluation test harnesse
 Command to run:
 
 ```bash
-dart run .agents/skills/pre-push-skill/tool/check_eval_commits.dart --remote=origin
+dart run evals/tool/check_eval_commits.dart --remote=origin
 ```
 
 *Note: You can optionally specify `--base-branch=<branch>` (defaults to `main`) or `--head=<ref>` (defaults to `HEAD`), or a different `--remote=<remote>` if comparing against a custom upstream repository or branch.*
