+++ packages/camera/camera_android_camerax/.agents/skills/pre-push-skill/evals/test_data/setup_success_native_test.dart
@@ -4,7 +4,7 @@
 
 import 'dart:io';
 
-import 'test_utils.dart';
+import '../../../../../evals/tool/test_utils.dart';
 
 void main() {
   ensureNotMainBranch();
