+++ packages/camera/camera_android_camerax/.agents/skills/pre-push-skill/evals/test_data/setup_user_facing_docs_change.dart
@@ -4,7 +4,7 @@
 
 import 'dart:io';
 
-import 'test_utils.dart';
+import '../../../../../evals/tool/test_utils.dart';
 
 void main() {
   ensureNotMainBranch();
