+++ packages/camera/camera_android_camerax/evals/tool/test_utils.dart
